Partially revert IPv4 related commits 72957f7 and 617f7ce.
This is a second attempt at running IPv6 only pod and service subnets,
with the repo now being mirrored on GitLab. BGP remains disabled, BPF
masquerading is also disabled since `autoDirectNodeRoutes` and native
routing for the IPv6 subnet _should_ take care of routing. If not, then
IPv6 masquerading will still have to be enabled; we'll see.